Fix resource leak and explorer restart bug
Icons were loaded each time active and inactive state toggled, now icons are allocated once, toggled, and destroyed on app exit. Also resolved the issue when explorer restarts we listen for TaskbarCreated event and reAdd() the icon. Before you would need to kill the app and relaunch for the icon to reappear.